refactor(broker): dedupe ttlSecondsFromQuery into internal/common
Same helper existed verbatim in emitter.go and completions/memory.go;
address review feedback on #2785 by sharing one exported copy.
